requirements.txt
asgiref==3.5.2
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
black==22.10.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
dj-database-url==1.0.0
Django==4.1.2
djangorestframework==3.14.0
drf-spectacular==0.24.2
executing==1.2.0
gunicorn==20.1.0
inflection==0.5.1
ipdb==0.13.9
ipython==8.6.0
jedi==0.18.1
jsonschema==4.17.0
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.10.1
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2-binary==2.9.5
pure-eval==0.2.2
pycodestyle==2.9.1
Pygments==2.13.0
pyrsistent==0.19.2
python-dotenv==0.21.0
pytz==2022.5
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.0
toml==0.10.2
tomli==2.0.1
traitlets==5.5.0
tzdata==2022.6
uritemplate==4.1.1
wcwidth==0.2.5
